+++ avc/cmd/avc/diff_current.go
@@ -8,8 +8,8 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/trevarix/agentic-vc/avc/internal/diff"
 	"github.com/spf13/cobra"
+	"github.com/trevarix/agentic-vc/avc/internal/diff"
 )
 
 var diffCurrentCmd = &cobra.Command{